### 2025-08-04 #8733 @editorialbot
[PRE REVIEW]: DynDom1D_Python: An Open-Source Python tool for the analysis of domain movements in proteins
**Submitting author:** @steven-hayward (<a href="http://orcid.org/0000-0001-6959-2604">steven hayward</a>)
**Repository:** https://github.com/steven-hayward/DynDom1D_Python
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
