### 2025-10-21 #8733 @steven-hayward
Hello, I understand that you have a large backlog of submissions so I was expecting some delay.  However, the young scientists who contributed to this work are understandably anxious to see it published. I think by now you should have an idea of how long we still have to wait. Please would you give us some idea when the review will commence?

Best regards
Steven Hayward
