### 2026-01-27 #8733 @Hughjmillard
Hi @rkurchin, I have been working with @steven-hayward on this project. Thank you for assigning yourself. Once we have made the required changes to the github page, should we create a new release of the software, or is saving the changes to the main branch enough?
Thanks,
Hugh
