### 2026-01-27 #8733 @rkurchin
No need for a new release at this point – we will very likely ask for one after any code/docs changes during the software review, but for things in the paper it's not necessary. Thanks for checking!
