### 2026-01-30 90bb0bc
Update paper.md
added new sections to paper md
