+++ paper.md
@@ -10,54 +10,71 @@ authors:
   - name: Jien Lim
     affiliation: 1
   - name: Hugh Millard
-    affiliation: 1
     orcid: 0009-0004-8963-1737
-  - name: Steven Hayward
     affiliation: 1
+  - name: Steven Hayward
     orcid: 0000-0001-6959-2604
     corresponding: true
+    affiliation: 1
 affiliations:
  - name: School of Computing Sciences, University of East Anglia, Norwich NR4 7TJ, U.K.
-   index: 1  
+   index: 1
+
 date: 1 August 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-DynDom1D_Python is an open-source Python implementation based on the original DynDom Fortran program for the analysis of domain movements in proteins. DynDom1D works on single protein chains and can be used when two structures of the same protein are available representing a conformational change. If appropriate it describes the conformational change in terms of the relative rotations of quasi-rigid regions called "dynamic domains" by way of hinge axes and hinge-bending residues. It improves over the original standalone version in a number of ways with an input file that requires only the protein data bank (PDB) accession codes and the chain identifiers of the two structures. The output has also been updated to produce a PyMol script file for display and animation of the domain movement.
+Here we introduce, DynDom1D_Python, an open-source Python implementation based on the original DynDom Fortran program for the analysis of domain movements in proteins. DynDom works on a single protein chain and can be used when two structures of the same protein are available representing a conformational change. If appropriate it describes the conformational change in terms of the relative rotation of quasi-rigid regions called "dynamic domains" by way of hinge axes (more precisely, interdomain screw axes) and hinge-bending residues (more precisely, interdomain bending residues). This new implementation improves on the original standalone version in a number of ways.
 
 # Statement of need
 
-Protein function usually involves interaction with other molecules, this interaction often effecting a change in conformation. Proteins are macromolecules that usually comprise many tens of thousands of atoms and conformational change occurs over a vast spatial scale. On the global scale is the domain movement where each domain comprises a significant portion of the whole protein. Protein domain movements form an important class of movements in proteins, perhaps the most important, and they are often engaged in all aspects of protein function, e.g. an enzyme trapping of a small substrate in its interdomain cleft. In principle, understanding how domain movements are controlled during function may be useful for drug design, e.g. by designing a drug to bind to a hinge-bending region in order to disrupt a functional hinge-bending movement.
+Protein function normally involves interaction with other molecules, this interaction often effecting a change in conformation. Proteins are macromolecules that usually comprise many tens of thousands of atoms and conformational change occurs over a vast spatial scale. On the global scale is the domain movement where each domain comprises a significant portion of the whole protein. Protein domain movements form an important class of movements in proteins, perhaps the most important, and they are often engaged in all aspects of protein function, e.g. an enzyme trapping of a small substrate in its interdomain cleft. In principle, understanding how domain movements are controlled during function may be useful for drug design, e.g. by designing a drug to bind to a hinge-bending region in order to disrupt a functional hinge-bending movement.
 
[…]
