### 2026-01-30 b1922f9
Update paper.bib
