+++ paper.bib
@@ -1,4 +1,5 @@
 @article{Hayward:1998,
+https://doi.org/10.1002/(SICI)1097-0134(19980201)30:2<144::AID-PROT4>3.0.CO;2-N,
 author = {Hayward, Steven and Berendsen, Herman J.C.},
 title = {Systematic analysis of domain motions in proteins from conformational change: New results on citrate synthase and T4 lysozyme},
 journal = {Proteins: Structure, Function, and Bioinformatics},
@@ -263,4 +264,94 @@ year = {1999}
   number={Oct},
   pages={2825--2830},
   year={2011}
+}
+@software{DomainFinder,
+  author={Konrad Hinsen},
+  title={DomainFinder},
+  url={https://github.com/khinsen/DomainFinder},
+  version = {2.0.5},
+  date = {2017-12-15},
+}
+
+
+@article{Gangwar2025,
+Author = {Gangwar, Shanti Pal and Yelshanskaya, Maria V. and Yen, Laura Y. and
+   Newton, Thomas P. and Sobolevsky, Alexander I.},
+Title = {Activation of kainate receptor GluK2-Neto2 complex},
+Journal = {NATURE STRUCTURAL \& MOLECULAR BIOLOGY},
+Year = {2025},
+Volume = {32},
+Number = {11},
+Pages = {2176-2184},
+Month = {NOV},
+DOI = {10.1038/s41594-025-01656-9},
+EarlyAccessDate = {AUG 2025},
+ISSN = {1545-9993},
+EISSN = {1545-9985},
+ResearcherID-Numbers = {Gangwar, Shanti/AAE-8163-2021
+   },
+ORCID-Numbers = {Newton, Thomas/0000-0002-2148-525X},
+Unique-ID = {WOS:001555771300001},
+}
+
+@article{Sotomayor2012,
+Author = {Sotomayor, Marcos and Weihofen, Wilhelm A. and Gaudet, Rachelle and
+   Corey, David P.},
+Title = {Structure of a force-conveying cadherin bond essential for inner-ear
+   mechanotransduction},
+Journal = {NATURE},
+Year = {2012},
+Volume = {492},
+Number = {7427},
+Pages = {128+},
+Month = {DEC 6},
+DOI = {10.1038/nature11590},
+ISSN = {0028-0836},
+EISSN = {1476-4687},
+ResearcherID-Numbers = {Sotomayor, Marcos/T-2411-2019
+   Gaudet, Rachelle/Y-3985-2019
+   },
+ORCID-Numbers = {Sotomayor, Marcos/0000-0002-3333-1805
+   Gaudet, Rachelle/0000-0002-9177-054X
+   Corey, David/0000-0003-4497-6016},
+Unique-ID = {WOS:000311893400059},
+}
+
+@article{HDAC62025,
+Author = {Iakovou, Georgios and Palmer, L. and Ganesan, A. and Kitao, Akio and
+   Laycock, Stephen D. and Hayward, Steven},
+Title = {Protein domain movement involved in binding of belinostat and HPOB as
+   inhibitors of histone deacetylase 6 (HDAC6): a hybrid
+   automated-interactive docking study},
+Journal = {JOURNAL OF COMPUTER-AIDED MOLECULAR DESIGN},
+Year = {2025},
+Volume = {39},
+Number = {1},
+Month = {JUL 15},
+DOI = {10.1007/s10822-025-00636-x},
+Article-Number = {52},
+ISSN = {0920-654X},
+EISSN = {1573-4951},
+ResearcherID-Numbers = {G, Dr. Annalakshmi/AFE-3431-2022
+   Kitao, Akio/B-8371-2008
+   },
[…]
