+++ paper.bib
@@ -1,5 +1,4 @@
 @article{Hayward:1998,
-https://doi.org/10.1002/(SICI)1097-0134(19980201)30:2<144::AID-PROT4>3.0.CO;2-N,
 author = {Hayward, Steven and Berendsen, Herman J.C.},
 title = {Systematic analysis of domain motions in proteins from conformational change: New results on citrate synthase and T4 lysozyme},
 journal = {Proteins: Structure, Function, and Bioinformatics},
