### 2026-01-30 5d243fe
Update paper.md missing sybol
