+++ .github/workflows/test-paper.yml
@@ -191,17 +191,17 @@ jobs:
         retention-days: 7
         if-no-files-found: ignore
         
-    - name: Comment on PR with PDF link (if PR)
-      uses: actions/github-script@v7
-      if: github.event_name == 'pull_request'
-      with:
-        script: |
-          github.rest.issues.createComment({
-            issue_number: context.issue.number,
-            owner: context.repo.owner,
-            repo: context.repo.repo,
-            body: 'JOSS paper compiled successfully! The generated PDF is available in the [workflow artifacts](https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }}).'
-          })
+    # - name: Comment on PR with PDF link (if PR)
+    #   uses: actions/github-script@v7
+    #   if: github.event_name == 'pull_request'
+    #   with:
+    #     script: |
+    #       github.rest.issues.createComment({
+    #         issue_number: context.issue.number,
+    #         owner: context.repo.owner,
+    #         repo: context.repo.repo,
+    #         body: 'JOSS paper compiled successfully! The generated PDF is available in the [workflow artifacts](https://github.com/${{ github.repository }}/actions/runs/${{ github.run_id }}).'
+    #       })
           
   validate-paper-content:
     runs-on: ubuntu-latest
