### 2026-01-31 #8733 @rkurchin
👋 @gpantel, @j-wags, and/or @xyzmjf, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
