### 2026-02-05 #8733 @Epivitae
> 👋 [@Epivitae](https://github.com/Epivitae), [@AnjaConev](https://github.com/AnjaConev), and/or [@braniii](https://github.com/braniii), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

sure, i can do this @rkurchin
