### 2026-02-22 482c638
fix command typo in README usage
