### 2026-02-22 685383e
Suggested install instructions for pymol
