### 2026-02-22 7488177
Fix inconsistent package name in __init__.py
