### 2026-02-22 a4a7009
Add GitHub Actions workflow to test the usage example
Adds .github/workflows/test-software.yml which:
1. Installs dependencies from src/DynDom1D_Python/requirements.txt
2. Saves the committed output files as a reference
3. Runs python main.py (the 4ake/2eck adenylate kinase example)
4. Diffs the freshly generated output against the saved reference

If the diff is non-empty the step fails, catching any unintended changes
to the program's output.  clustering_logs/ is excluded from the diff
because it contains timestamped files that change on every run.

Co-Authored-By: Claude Sonnet 4.6 <noreply
[…]
