### 2026-02-22 8f96c67
Add weekly scheduled run to CI workflow
Runs every Monday at 06:00 UTC to catch breakage from upstream
dependency updates even when the repo itself is quiet.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
