### 2026-02-22 fa92fb3
Port dependency comments from requirements.txt into pyproject.toml
TOML supports # comments, so the category labels and optional-dependency
note from the old requirements.txt are preserved here.

Co-Authored-By: Claude Sonnet 4.6 <noreply@anthropic.com>
