### 2026-02-22 a4a28e1
fully transfer commented out deps from requirements.txt
