### 2026-02-22 #9938 @j-wags
This looks like a valuable rewrite of a widely used tool. My main concern is the lack of tests and testing infrastructure. I've opened a PR to add a GitHub Actions workflow that runs the usage example and checks the output, but I think more testing than that will be needed for this submission (I was waffling on objecting over the lack of tests until I reread the first sentence of the [JOSS Guiding Principles](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#guiding-principles) and it mentioned tests and continuous integration). I'd be happy to re-review this later if more tests are added, and I hope my comments and PRs can be helpful in polishing some things and getting a continuous integration testing workflow started.

I've categorized each of my comments as (blocking) or (not blocking). The (not blocking) comments are suggestions that aren't in the way of me approving this submission. However the (blocking) ones are blocking my approval of this submission, and I've included checkboxes by them to be clear about when they're resolved. I've opened PRs to fix some of the (blocking) issues. If those get merged (or the issues are addressed some other way) I'll check the corresponding box.

Authorship:

* (not blocking) Jien Lim is listed as an author but doesn’t appear in the git contributor history. I assume this is for substantial non-code contributions or work done offline/before the code was open-sourced but just double checking.

Installation:

* [x] (blocking) There are some typos in the usage example that prevent it from working out of the box. I propose fixes in https://github.com/steven-hayward/DynDom1D_Python/pull/2
* (not blocking) PyMol is a fairly important (yet technically optional) dependency for using this software. I propose adding some information about getting it in https://github.com/steven-hayward/DynDom1D_Python/pull/3
* (not blocking) Dependencies are defined identically in both `requirements.txt` and `pyproject.toml`. It seems like it’ll be easy for them to fall out of sync with each other. I propose a consolidation of them (which updates the docs accordingly) in https://github.com/steven-hayward/DynDom1D_Python/pull/6
* (not blocking) The package loses a lot of downstream value by not being installable, and needing to run `main.py` in a repo directory.

Running:

* [x] (blocking) When I ran the provided example, I found some differences in the `_info` file (all numerical outputs were identical, but there’s some different colors and additional logging output in the new version). I update these files and implemented automated testing against their contents in [https://github.com/steven-hayward/DynDom1D_Python/pull/5](https://github.com/steven-hayward/DynDom1D_Python/pull/5)
* (not blocking) I was easily able to modify the provided inputs and run with two calmodulin structures (1AHR and 1A29), and the software worked well.

Development history and open source practice:

* Open development/Collaborative effort notes:
  * (not blocking) The majority of code in the repo appears to have been uploaded on June 18 in this commit: [https://github.com/steven-hayward/DynDom1D_Python/commit/a6b01871c90fc47bc453ef359b57168750a24551](https://github.com/steven-hayward/DynDom1D_Python/commit/a6b01871c90fc47bc453ef359b57168750a24551) and then moderately changed over the next weeks, and it's been mostly undeveloped since.
  * (not blocking) There’s no evidence of external engagement (issues/PRs from non authors)
* Good practices notes:
  * (not blocking) The repo has a license file and strong documentation in the README. It doesn’t appear to have tests/a change verification process or a contributor guide. The support pathway can be assumed to be the issue tracker, and the README has a nice FAQ that will probably head off a majority of user issues.

Documentation:

* Automated tests notes:
  * [x] (blocking) Neither tests nor automation that runs tests appears to exist. Some level of testing will be necessary to catch upstream breaks and to give contributors a way to propose changes. I propose adding an automated test that runs the example and diffs the output against the stored version to ensure they’re identical (This should help the maintainers: if a PR would modify the output, then that will be reflected in needing to change the reference output to get tests passing). I propose adding testing in https://github.com/steven-hayward/DynDom1D_Python/pull/5
* Community guidelines:
  * (not blocking) These don’t appear to exist.
* [x] (blocking) The pair of PDB codes suggested in the auto-downloading usage example (1hng 2hng) causes an error if chain A of both is requested, and only works if the input is (1hng chain B) and (2hng chain A). Running a clustalw alignment shows low sequence identity between these two. If these PDB codes continue to be used as an example, the instructions should either list which chain to use and highlight that this is a difficult case with low sequence identity, or suggest a more similar pair of PDB structures.

Software paper:

* (not blocking) I really appreciate the history/context provided in the paper. It contains a very strong explanation of why the python rewrite of this tool will be valuable.
* (not blocking) The JOSS editor bot is complaining about some references - The authors should check those and either fix the DOIs or attest that the bibliography is correct despite the bot complaints.
  * https://github.com/openjournals/joss-reviews/issues/9938#issuecomment-3863000966
