### 2026-02-24 #9938 @AnjaConev
Seems like a very useful and welcome addition to the open source python packages for analysis of protein conformational change. Here are a few major comments:

**Development**

- Open development (not blocking)
    - most of the core code was uploaded in a single commit so it seems that the majority of development was done externally and imported to GitHub at a final stage. However, since the initial commit there has been engagement and further commits, mainly by the two main authors, so hopefully this good practice will continue.
- Development good practices (not blocking)
    - missing documentation and tests

**Functionality**
- Installation (not blocking)
    - Git clone instruction has a typo: yourusername should be replaced with the real repo name
    - Requirements.txt is not located at the root for the repo, so wrong path provided
    - When I fixed those two, installed smoothly
- Functionality (blocking)
    - Typos in instructions might make it difficult for beginner users to run, for example `python ./main.py` can not run cause main.py is not in data or root directories, better specify the full path or make your package pip installable
    - When running main.py I am getting the following error:
   ```
        * File "/Users/anjaconev/DynDom1D_Python/src/DynDom1D_Python/main.py", line 348, in determine_domains_screw_axis
        *     self.log_fixed_domain_coordinates("After fixed-domain transformation")
        *     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        * AttributeError: 'Engine' object has no attribute 'log_fixed_domain_coordinates'

I will continue the review in the following days, but I am concerned about the current level of functionality so want to give the authors the time/notice to fix these bugs.
