### 2026-02-26 #9938 @steven-hayward
@AnjaConev
Many thanks for agreeing to review DynDom1D_Python. We would like to see if we can reproduce the run-time error you found. Can you inform us of the PDB codes and the chain ID's of the PDB structures you used?

Best regards
Steven Hayward
