### 2026-02-26 #9938 @AnjaConev
> [@AnjaConev](https://github.com/AnjaConev) Many thanks for agreeing to review DynDom1D_Python. We would like to see if we can reproduce the run-time error you found. Can you inform us of the PDB codes and the chain ID's of the PDB structures you used?
>
> Best regards Steven Hayward

I just raised an issue here with a detailed description: https://github.com/steven-hayward/DynDom1D_Python/issues/7
I hope it is detailed enough, but please reply in the issue if you are not able to reproduce.
