### 2026-03-08 49675b5
add gitignore + update readme
