### 2026-03-08 f213878
update filemngr and main py fixing sequence identity and bending residue bugs
