### 2026-03-08 cccaba1
updating test-software so outputs are compared with an output saved as a reference
