+++ .github/workflows/test-software.yml
@@ -1,5 +1,4 @@
 name: Test Software
-
 # Run on every push and pull request to main/master, weekly to catch breakage
 # from upstream dependency updates, and allow manual triggering.
 on:
@@ -15,7 +14,6 @@ jobs:
   run-example:
     runs-on: ubuntu-latest
     name: Run usage example and verify output
-
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
@@ -28,11 +26,6 @@ jobs:
       - name: Install dependencies
         # requirements.txt lives inside the package directory, not at the repo root.
         run: pip install -r src/DynDom1D_Python/requirements.txt
-
-      - name: Save reference outputs
-        # The output files already committed in the repo are the reference.
-        # Copy them now, before running the script, so we can compare afterward.
-        run: cp -r src/DynDom1D_Python/output/4ake_A_2eck_B reference_output
 
       - name: Run the example
         # main.py reads data/command.txt (pre-configured for the 4ake chain A /
@@ -47,12 +40,16 @@ jobs:
         # clustering_logs/ is excluded because it contains timestamped log files
         # that are intentionally different on every run.
         #
+        # The reference outputs are stored in reference_output/4ake_A_2eck_B/ at
+        # the repo root and are committed once as a known-good snapshot. They are
+        # only updated deliberately — never overwritten automatically by CI.
+        #
         # If this step fails, the code is producing different output than the
-        # committed reference.  To update the reference, run the example locally
-        # (from src/DynDom1D_Python/) and commit the updated files in
-        # src/DynDom1D_Python/output/4ake_A_2eck_B/.
+        # committed reference. To update the reference, run the example locally
+        # (from src/DynDom1D_Python/) and copy the results:
+        #   cp -r src/DynDom1D_Python/output/4ake_A_2eck_B reference_output/
         run: |
           diff -r \
             --exclude="clustering_logs" \
             src/DynDom1D_Python/output/4ake_A_2eck_B \
-            reference_output
+            reference_output/4ake_A_2eck_B
