### 2026-03-08 1db8a92
Changed w5 info to just be info (not all files have window of 5) and corrected encoding bug causing tests to fail
