### 2026-03-09 3c83455
Update README.md
Change command txt example
