### 2026-03-09 #9938 @stevenjhayward61-commits
@j-wags @AnjaConev Thank you for your efforts in reviewing this project.  Here are our responses:

Authorship: Jien Lim worked outside of Github two years ago on a short internship to develop the code. Unfortunately Jien could not complete the project. Hugh Millard took over the project, finished it after a lot of testing and put it on Gitub.

Installation: We believe the blocking comments have been dealt with.

Documentation: In the README the recommended PDB files for testing were 1hng and 2hng. These contain completely different proteins and although we have wracked our brains we do not know the reason for this error. This has now been corrected to 4ake and 2eck. We have added the automated test/check. Thank you for this excellent suggestion.

Running/Functionality: Thank you for picking up the error regarding the sequence identity threshold. This has now been corrected. In the course of further testing we picked up another bug: in the bending residue numbering in the pml file mismatched that in the info file. This was due to it not correcting to the centre of the sliding window in the pml file. Corrected.

Software/Paper: We did not get an error from the bot concerning the references. We can look again if we need to make further changes to the paper.

Best wishes
Steven Hayward
