### 2026-03-19 9c72820
Update README with clearer usage instructions
Clarified instructions for preparing input files and running the analysis.
