### 2026-03-20 #9938 @stevenjhayward61-commits
@AnjaConev @j-wags @rkurchin

We have responded to the latest issues in:
https://github.com/steven-hayward/DynDom1D_Python/issues/7

We are looking forward to hearing more comments if there are any.
