### 2026-04-07 #9938 @j-wags
Thanks for responding to the issues and apologies for the delays on my end!

My concerns have been resolved except one. I'd like to check with @rkurchin on JOSS's expectations around testing. We've added a GitHub Actions test run of the README example but there are no tests beyond that. This is my first JOSS review, so I searched around other Python submissions and saw that nearly all of them had more substantial tests. Is the current state of testing a blocker to this submission?
