### 2026-04-16 e29fd84
add testing command files
