+++ src/DynDom1D_Python/data/1cdllA_1cllA_command.txt
@@ -1,0 +1,17 @@
+# DIRECTORY WHERE PDB FILES ARE LOCATED.
+input_path=data/pdb
+#
+# DIRECTORY WHERE OUTPUT WILL BE PUT.
+output_path=output
+#
+# WRITE FILENAME FOR 1ST CHAIN
+filename1=1CDL
+#
+# ONE LETTER CODE FOR 1ST CHAIN ID IF IT EXISTS, OTHERWISE LEAVE BLANK
+chain1id=A
+#
+# WRITE FILENAME FOR 2ND CHAIN, MAY BE SAME AS FIRST
+filename2=1CLL
+#
+# ONE LETTER CODE FOR 2ND CHAIN ID
+chain2id=A
