### 2026-04-16 #9938 @steven-hayward
We have now implemented further tests in the test-software.yml. These include the original test on adenylate kinase in the README and additional tests on citrate synthase, alcohol dehydrogenase, maltodextrin binding protein and calmodulin.

We hope that this latest improvement will be sufficient and look forward to the submission going on to the final stage.
