### 2026-07-20 3607516
Update paper.md
post review
