+++ paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-Here we introduce, DynDom1D_Python, an open-source Python implementation based on the original DynDom Fortran program for the analysis of domain movements in proteins. DynDom works on a single protein chain and can be used when two structures of the same protein are available representing a conformational change. If appropriate it describes the conformational change in terms of the relative rotation of quasi-rigid regions called "dynamic domains" by way of hinge axes (more precisely, interdomain screw axes) and hinge-bending residues (more precisely, interdomain bending residues). This new implementation improves on the original standalone version in a number of ways.
+Here we introduce DynDom1D_Python, an open-source Python implementation based on the original DynDom Fortran program for the analysis of domain movements in proteins. DynDom works on a single protein chain and can be used when two structures of the same protein are available representing a conformational change. If appropriate, it describes the conformational change in terms of the relative rotation of quasi-rigid regions called "dynamic domains" by way of hinge axes (more precisely, interdomain screw axes) and hinge-bending residues (more precisely, interdomain bending residues). This new implementation improves on the original standalone version in a number of ways.
 
 # Statement of need
 
@@ -36,7 +36,7 @@ The original Fortran implementation of the DynDom program [@Hayward:1998;@Lee:20
 
 # State of the field
 
-There are two main approaches to the task of determining dynamic domains and hinge axes from a pair of structures. The DynDom tools and the DomainFinder tool [@Hinsen:1999] cluster parameters relating to rigid body motions to determine dynamic domains. The HingeFind approach of Wriggers and Schulten [@Wriggers:1997] determines domains using an adaptive best-fitting procedure and is run from within X-PLOR or the VMD environment. There are distinct differences between all three tools. DomainFinder [@DomainFinder] performs normal mode analysis on a single structure to determine its modes of movement and does not directly accept two structure files. HingeFind determines the rotation axis that best describes the domain movement, whereas DynDom and DomainFinder determine the screw axis. DynDom1D is unique in being able to determine hinge-bending residues as part of an overall coherent process. As stated above hinge-bending regions are potential binding sites for drug molecules.
+There are two main approaches to the task of determining dynamic domains and hinge axes from a pair of structures. The DynDom tools and the DomainFinder tool [@Hinsen:1999] cluster parameters relating to rigid-body motions to determine dynamic domains. The HingeFind approach of Wriggers and Schulten [@Wriggers:1997] determines domains using an adaptive best-fitting procedure and is run from within X-PLOR or the VMD environment. There are distinct differences between all three tools. DomainFinder [@DomainFinder] performs normal mode analysis on a single structure to determine its modes of movement and does not directly accept two structure files. HingeFind determines the rotation axis that best describes the domain movement, whereas DynDom and DomainFinder determine the screw axis. DynDom1D is unique in being able to determine hinge-bending residues as part of an overall coherent process. As stated above, hinge-bending regions are potential binding sites for drug molecules.
 
 
 # Research Impact Statement
@@ -45,7 +45,7 @@ The Fortran version of DynDom1D was released in 1998 and as part of CCP4 from 19
 
 # Method
 
[…]
