### 2026-07-20 eb84145
Update paper.md
