### 2026-07-20 626bd70
Update paper.md
