### 2026-07-20 f563916
Update paper.md
