+++ paper.md
@@ -60,7 +60,7 @@ The basic idea behind the DynDom approach is that domains can be recognised as r
 11. Assign mechanical hinges and percentage "closure".
 12. Output results including PDB files for display with PyMol using a PyMol script file (.pml file).
 
-Further details on the basic methodology can be found in the original DynDom publications [@Hayward:1998;@Lee:2002]. An important feature is its ability to assign "mechanical hinges". These are hinge-bending regions that are close to (within 5.5 &angst) the hinge axis [@Hayward:1999]. In order to properly control a domain movement, two or more mechanical hinges are separated in space determining the direction and location of the screw axis. Often such arrangements create a "closure" motion whereby the interdomain cleft narrows or widens depending on the direction of rotation. DynDom1D_Python assigns mechanical hinges, and the percentage closure motion (0% closure implies 100% twisting) and also gives a point on the screw axis and its direction in terms of the unit vector.  
+Further details on the basic methodology can be found in the original DynDom publications [@Hayward:1998;@Lee:2002]. An important feature is its ability to assign "mechanical hinges". These are hinge-bending regions that are close to (within 5.5 &#x00C5;) the hinge axis [@Hayward:1999]. In order to properly control a domain movement, two or more mechanical hinges are separated in space determining the direction and location of the screw axis. Often such arrangements create a "closure" motion whereby the interdomain cleft narrows or widens depending on the direction of rotation. DynDom1D_Python assigns mechanical hinges, and the percentage closure motion (0% closure implies 100% twisting) and also gives a point on the screw axis and its direction in terms of the unit vector.  
 
 # Software design
 
