### 2026-07-20 ed25245
Update paper.bib
Bib update
