+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Hayward:1998,
-author = {Hayward, Steven and Berendsen, Herman J.C.},
+author = {Hayward, S. and Berendsen, H.J.C.},
 title = {Systematic analysis of domain motions in proteins from conformational change: New results on citrate synthase and T4 lysozyme},
 journal = {Proteins: Structure, Function, and Bioinformatics},
 volume = {30},
@@ -15,9 +15,9 @@ year = {1998}
 
 
 @article{Hinsen:1999,
-Author = {Hinsen, K and Thomas, A and Field, MJ},
+Author = {Hinsen, K and Thomas, A.and Field, M.J.},
 Title = {Analysis of domain motions in large proteins},
-Journal = {PROTEINS-STRUCTURE FUNCTION AND BIOINFORMATICS},
+Journal = {Proteins-Structure Function and Bioinformatics},
 Year = {1999},
 Volume = {34},
 Number = {3},
@@ -34,10 +34,10 @@ Unique-ID = {WOS:000078376300009},
 
 
 @article{Wriggers:1997,
-Author = {Wriggers, W and Schulten, K},
+Author = {Wriggers, W. and Schulten, K.},
 Title = {Protein domain movements: Detection of rigid domains and visualization
    of hinges in comparisons of atomic coordinates},
-Journal = {PROTEINS-STRUCTURE FUNCTION AND BIOINFORMATICS},
+Journal = {Proteins-Structure Function and Bioinformatics},
 Year = {1997},
 Volume = {29},
 Number = {1},
@@ -53,10 +53,10 @@ Unique-ID = {WOS:A1997XU73100001},
 
 
 @article{Veevers:2019,
-Author = {Veevers, Ruth and Hayward, Steven},
+Author = {Veevers, R. and Hayward, S.},
 Title = {Methodological improvements for the analysis of domain movements in
    large biomolecular complexes},
-Journal = {BIOPHYSICS AND PHYSICOBIOLOGY},
+Journal = {Biophysics and Physicobiology},
 Year = {2019},
 Volume = {16},
 Pages = {328-336},
@@ -69,11 +69,11 @@ Unique-ID = {WOS:000503189000038},
 
 
 @article{Poornam:2009,
-Author = {Poornam, Guru Prasad and Matsumoto, Atsushi and Ishida, Hisashi and
-   Hayward, Steven},
+Author = {Poornam, G. P. and Matsumoto, A. and Ishida, H. and
+   Hayward, S.},
 Title = {A method for the analysis of domain movements in large biomolecular
    complexes},
-Journal = {PROTEINS-STRUCTURE FUNCTION AND BIOINFORMATICS},
+Journal = {Proteins-Structure Function and Bioinformatics},
 Year = {2009},
 Volume = {76},
 Number = {1},
@@ -87,10 +87,10 @@ Unique-ID = {WOS:000266370100016},
 }
 
 @article{Lee:2002,
-Author = {Hayward, S and Lee, RA},
+Author = {Hayward, S and Lee, R.A.},
 Title = {Improvements in the analysis of domain motions in proteins from
    conformational change: DynDom version 1.50},
-Journal = {JOURNAL OF MOLECULAR GRAPHICS \& MODELLING},
+Journal = {Journal Of Molecular Graphics \& Modelling},
 Year = {2002},
 Volume = {21},
 Number = {3},
@@ -106,9 +106,9 @@ Unique-ID = {WOS:000179393700003},
 
 
 @article{Lee:2003,
-Author = {Lee, RA and Razaz, M and Hayward, S},
+Author = {Lee, R.A. and Razaz, M. and Hayward, S},
[…]
