### 2026-07-20 8aa598d
Update paper.bib
update b ib
