+++ paper.bib
@@ -42,6 +42,7 @@ Year = {1997},
 Volume = {29},
 Number = {1},
 Pages = {1-14},
+DOI = {https://doi.org/10.1002/(SICI)1097-0134(199709)29:1%3C1::AID-PROT1%3E3.0.CO;2-J},
 Month = {SEP},
 ISSN = {0887-3585},
 EISSN = {1097-0134},
@@ -60,7 +61,7 @@ Journal = {Biophysics and Physicobiology},
 Year = {2019},
 Volume = {16},
 Pages = {328-336},
-DOI = {10.2142/biophysico.16.0\_328},
+DOI = {https://doi.org/10.2142/biophysico.16.0_328},
 EISSN = {2189-4779},
 ORCID-Numbers = {hayward, steven/0000-0001-6959-2604
    Veevers, Ruth/0000-0001-7599-5725},
@@ -298,7 +299,7 @@ Author = {Sotomayor, Marcos and Weihofen, Wilhelm A. and Gaudet, Rachelle and
    Corey, David P.},
 Title = {Structure of a force-conveying cadherin bond essential for inner-ear
    mechanotransduction},
-Journal = {NATURE},
+Journal = {Nature},
 Year = {2012},
 Volume = {492},
 Number = {7427},
