### 2026-07-20 1faafa0
Update paper.bib
