+++ paper.bib
@@ -15,7 +15,7 @@ year = {1998}
 
 
 @article{Hinsen:1999,
-Author = {Hinsen, K and Thomas, A.and Field, M.J.},
+Author = {Hinsen, K. and Thomas, A. and Field, M.J.},
 Title = {Analysis of domain motions in large proteins},
 Journal = {Proteins-Structure Function and Bioinformatics},
 Year = {1999},
@@ -88,7 +88,7 @@ Unique-ID = {WOS:000266370100016},
 }
 
 @article{Lee:2002,
-Author = {Hayward, S and Lee, R.A.},
+Author = {Hayward, S. and Lee, R.A.},
 Title = {Improvements in the analysis of domain motions in proteins from
    conformational change: DynDom version 1.50},
 Journal = {Journal Of Molecular Graphics \& Modelling},
@@ -266,7 +266,7 @@ year = {1999}
   year={2011}
 }
 @software{DomainFinder,
-  author={Konrad Hinsen},
+  author={K.Hinsen},
   title={DomainFinder},
   url={https://github.com/khinsen/DomainFinder},
   version = {2.0.5},
