### 2026-07-24 bffe63e
Update paper.bib
Final minor updates
