+++ paper.bib
@@ -6,7 +6,7 @@ volume = {30},
 number = {2},
 pages = {144-154},
 keywords = {hinge bending, X-ray conformers, Chasles' theorem},
-doi = {https://doi.org/10.1002/(SICI)1097-0134(19980201)30:2<144::AID-PROT4>3.0.CO;2-N},
+doi = {10.1002/(SICI)1097-0134(19980201)30:2<144::AID-PROT4>3.0.CO;2-N},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/%28SICI%291097-0134%2819980201%2930%3A2%3C144%3A%3AAID-PROT4%3E3.0.CO%3B2-N},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/%28SICI%291097-0134%2819980201%2930%3A2%3C144%3A%3AAID-PROT4%3E3.0.CO%3B2-N},
 abstract = {Abstract Methods developed originally to analyze domain motions from simulation [Proteins 27:425–437, 1997] are adapted and extended for the analysis of X-ray conformers and for proteins with more than two domains. The method can be applied as an automatic procedure to any case where more than one conformation is available. The basis of the methodology is that domains can be recognized from the difference in the parameters governing their quasi-rigid body motion, and in particular their rotation vectors. A clustering algorithm is used to determine clusters of rotation vectors corresponding to main-chain segments that form possible dynamic domains. Domains are accepted for further analysis on the basis of a ratio of interdomain to intradomain fluctuation, and Chasles' theorem is used to determine interdomain screw axes. Finally residues involved in the interdomain motion are identified. The methodology is tested on citrate synthase and the M6I mutant of T4 lysozyme. In both cases new aspects to their conformational change are revealed, as are individual residues intimately involved in their dynamics. For citrate synthase the beta sheet is identified to be part of the hinging mechanism. In the case of T4 lysozyme, one of the four transitions in the pathway from the closed to the open conformation, furnished four dynamic domains rather than the expected two. This result indicates that the number of dynamic domains a protein possesses may not be a constant of the motion. Proteins 30:144–154, 1998. © 1998 Wiley-Liss, Inc.},
@@ -42,7 +42,7 @@ Year = {1997},
 Volume = {29},
 Number = {1},
 Pages = {1-14},
-DOI = {https://doi.org/10.1002/(SICI)1097-0134(199709)29:1%3C1::AID-PROT1%3E3.0.CO;2-J},
+DOI = {10.1002/(SICI)1097-0134(199709)29:1%3C1::AID-PROT1%3E3.0.CO;2-J},
 Month = {SEP},
 ISSN = {0887-3585},
 EISSN = {1097-0134},
@@ -61,7 +61,7 @@ Journal = {Biophysics and Physicobiology},
 Year = {2019},
 Volume = {16},
 Pages = {328-336},
-DOI = {https://doi.org/10.2142/biophysico.16.0_328},
+DOI = {10.2142/biophysico.16.0_328},
 EISSN = {2189-4779},
 ORCID-Numbers = {hayward, steven/0000-0001-6959-2604
    Veevers, Ruth/0000-0001-7599-5725},
@@ -225,7 +225,7 @@ Unique-ID = {WOS:001004084500003},
 }
 @article{Wojdyr:2022, 
 doi = {10.21105/joss.04200}, 
-url = {https://doi.org/10.21105/joss.04200},
+url = {10.21105/joss.04200},
 year = {2022}, 
 publisher = {The Open Journal}, 
 volume = {7}, 
@@ -250,14 +250,14 @@ volume = {36},
 number = {4},
 pages = {425-435},
[…]
