+++ paper.bib
@@ -257,7 +257,7 @@ abstract = {Abstract With the use of a recently developed method, twenty-four pr
 year = {1999}
 }
 @article{pedregosa2011scikit:2011,
-  title={Scikit-learn: Machine learning in \MakeUppercase{P}ython},
+  title={Scikit-learn: Machine learning in {P}ython},
   author={Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and others},
   journal={Journal of machine learning research},
   volume={12},
