### 2026-07-29 3f4d2f2
Update paper.md
U.K. to United Kingdom
