### 2026-07-31 00f7916
added contributing.md
