### 2026-07-31 a25c0c3
Add contributing guidelines to README
Added a contributing section to the README.
